requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface

groq

faiss-cpu
sentence-transformers
pymupdf

streamlit

python-dotenv
fpdf2
tavily-python